+++ front_matter/deno.json
@@ -7,6 +7,7 @@
     "./json": "./json.ts",
     "./test": "./test.ts",
     "./toml": "./toml.ts",
-    "./yaml": "./yaml.ts"
+    "./yaml": "./yaml.ts",
+    "./types": "./types.ts"
   }
 }
