+++ front_matter/json.ts
@@ -1,10 +1,9 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
-import {
-  createExtractor,
-  type Extractor,
-  type Parser,
-} from "./_create_extractor.ts";
+import { createExtractor, type Parser } from "./_create_extractor.ts";
+import type { Extractor } from "./types.ts";
+
+export type { Extractor };
 
 /**
  * Extracts and parses {@link https://www.json.org/ | JSON } from the metadata
